Fix a crash caused by our new set.union signatures.
The 'instance' attribute of a TypeParameterInstance can be set to a
non-instance object. We weren't accounting for this possibility in output.py,
hence the crash when trying to output a type containing a TypeParameterInstance
with its instance set to an abstract.function.Signature.

After this change, the crashes involving the set.union signatures are gone.
They've been replaced by some slightly questionable pytype findings, but that's
probably fine, given how rare this case is.

PiperOrigin-RevId: 573250340