--bind-decorated-methods: Don't treat SimpleFunction as a full-fledged method.
--bind-decorated-methods works by analyzing the non-decorated version of a
method when the decorated one has a non-method type. This didn't work properly
when the decorated method was a SimpleFunction because pytype considered a
SimpleFunction a method even though it doesn't have a body to analyze.

I had to tweak two paramspec tests that were accidentally relying on pytype not
checking the type of 'self' for a SimpleFunction.

PiperOrigin-RevId: 596667800