Enable access to dependencies via import map.
Uses the target label given via the project and the existing
pytype API to locate the `imports_info` path for the file
being type-checked.

This file path is then passed in to the pytype configuration options
prior to type checking being run, allowing the map to be used to locate
the (already generated) ASTs for dependent modules.

Adds a test case which verifies that the correct file path
corresponding to the target label is used during configuration.
We opted not to use a more complex test actually verifying the
loading and use of serialized ASTs because of the following:
- The parsing of the file and loading of ASTs is a pytype implementation
  detail.
- Testing the loading of serialized ASTs would turn our test into a change
  detector for the internal serialization format used by pytype.
- Checking that the file is attempted to be used and has the expected path
  form is sufficient for testing the code external to pytype.

#pytype-in-cider

PiperOrigin-RevId: 601840387