Start a pytype/rewrite subdirectory.
This is mainly to get the directory set up, but I also added a variables.py
with Variable and Binding classes, and a Condition superclass. Everything is a
frozen dataclass so that, down the line, we have to make copies of things to
update a block's state and can't accidentally modify another block's state at
the same time.

PiperOrigin-RevId: 601892043