Remove 'deep' and 'store_all_calls' options.
I want rewrite/ to use its own versions of analyze.check_types and
analyze.infer_types as entrypoints, but the 'deep' option is bubbled all the
way up to io.py, so rewrite/ would have to support it.

I can't remember ever doing anything useful with 'deep', so we might as well
remove it. After 'deep' is removed, 'store_all_calls' isn't used anywhere
except in tools/xref/indexer/, and I verified that the indexer tests pass
without it, so it can be removed as well.

PiperOrigin-RevId: 601964092