Remove duplicate pytype inputs and outputs.
A lot of information was being passed to or returned from various pytype
functions multiple times - e.g., io.AnalysisResult contained both an 'errorlog'
field and a 'context' field that had the errorlog as an attribute. This change
deduplicates as much as possible.

I also removed a 'generate_unknowns' option to context.Context, since it was
simply a combination of the 'check' and 'protocols' options, and a
'show_library_calls' option to analyze.infer_types that was used only in tests.

PiperOrigin-RevId: 602804753