Fix module resolution bug in load_pytd.
See https://github.com/python/typeshed/pull/11329 for a case in which the old
approach failed.

Hypothetically, this new approach could get quite expensive, with
LookupExternalTypes called over and over, but I ran typeshed's pytype_test a
few times with and without the change and noticed no difference in runtime.

PiperOrigin-RevId: 603435155