Define a set of opaque base classes for pytype objects.
These classes are "opaque" in the sense that they do not expose details of
pytype's internal representation of abstract objects and typegraph elements.
The intended use is to decouple as much of pytype's code as possible from these
internal details, so that it can be shared between the current codebase and the
rewrite.

PiperOrigin-RevId: 617290604