Rename errorlog classes.
Since we export `ErrorLog` as a return type from `pytype.io` functions, callers
often want to create an empty `ErrorLog` object when catching a pytype
exception, simply so that the return value can be properly typed and/or passed
to another function that expects an `ErrorLog` object.

We therefore rename `ErrorLogBase` to `ErrorLog`, since it satisfies the
contract external callers want from the return type, and the internal
`ErrorLog` to `VmErrorLog`. In particular, the `ErrorLog` constructor can now
be called without needing a pretty-printer, since it does not need to add new
errors.

PiperOrigin-RevId: 619004930