Update pytype_runner.py - change escape_ninja_path to escape all the ninja path escape-needing characters (#1608)
This fixes https://github.com/google/pytype/issues/499. Unfortunately, I was not able to get pytype running on my machine, so this commit is just based on what seems like the right idea to me, and hopefully it can be verified to be correct by a maintainer and then merged.

Resolves #1608

PiperOrigin-RevId: 623037901