rewrite: Unpack *args and **args when mapping against a signature.
Largely ported from abstract.function.Args.simplify, with a few changes because
we don't handle indefinite iterables well yet, and refactored a bit to be
easier to read and type check.

Not tested for now because it also needs changes to LIST_EXTEND and a few other
places to handle unpacking of `Any` better, and this change was getting long
enough already. Added a skipped test that works in the execution phase, but not
the analyse-with-fake-args phase.

PiperOrigin-RevId: 625859869