rewrite: add a class transformation mechanism to overlays.overlays.
With this, overlays that need to transform classes can simply write a
transformation function and register it via a decorator. It's a little
less flexible than having complete control over the metaclass's __new__
function, but it's a lot more convenient to work with (no need to pack the
arguments to __new__ into variables and then unpack them in the overlay), and
we can reuse the same transformation mechanism for imported classes.

For now, I've implemented just enough for EnumClassTransform to work. My next
step will be to tackle dataclasses and typing.dataclass_transform, which should
help me figure out what other "hooks" we need for class transformation.

PiperOrigin-RevId: 627892423