Update stubs for 3.12
Add stubs that exist in typeshed and are gated for `sys.version_info >= (3, 12)`.

This is a best effort. The existing .pytd files looks quite different to the .pyi files in typeshed. I only looked for version-gated definitions in typeshed and tried to guess if they would be beneficial to have in pytype.

Especially important is `TypeAliasType`. Without that, most imports of stdlib modules fail, because they end up transitively importing that class.

Failing tests: before=1274, after=490

PiperOrigin-RevId: 648737180