Add mapping for CALL_INTRINSIC_{1,2} argument
Python 3.12 adds new opcodes CALL_INTRINSIC_1 and CALL_INTRINSIC_2. The opcode argument is the name of the internal function to call.

This adds a mapping to pycnite, so `op.argval` contains the string representation of the internal function name, e.g. for CALL_INTRINSIC_1 if op.arg is 1 op.argval is "INTRINSIC_PRINT".

PiperOrigin-RevId: 649409008