Undo FOR_ITER/END_FOR fix for 3.12
Undo https://github.com/google/pytype/commit/72197ea5dd6ebd0a2e28456c26f434b2ebac0b40 and explain why in comments.

In 3.12+, FOR_ITER still pops the iterator off the stack, but now conditionally. See https://github.com/python/cpython/issues/121399.

Pytype doesn't actually execute the loop, so we need to handle this differently. We always pop the iterator after FOR_ITER, same as in <=3.11 and make END_FOR is a no-op. Since we don't execute the loop, we never have a situation where at the end of the loop the top of the stack is `[iter, iter()]`, so the double-pop of END_FOR is not needed.

I really hope this is correct. We fix more tests with this, so it can't be totally wrong.

Failing tests: before=489, after=432

PiperOrigin-RevId: 650618420