Treat setting fast local to null as a delete (for 3.12)
Python internally distinguishes between locals (variables stored in frame.f_locals) and "fast locals" (variables stored essentially at the bottom of the stack).

This typically doesn't matter because before locals are accessed, fast locals are copied into locals using `PyFrame_FastToLocals`[1].

But it _does_ matter when setting fast locals to NULL. Because setting a fast local to null means deleting the local.[2]

Not sure why this hasn't been an issues n 3.11 already. But 3.12 uses this in a few more places. E.g. list comprehensions are now inlined and the iteration variable is a fast local and set to null in the end. This change is needed to tell Pytype that the iteration variable does not exist anymore and should not be included in .pyi files, etc.

[1]: https://github.com/python/cpython/blo[]d1f15712d243e03e50332abb86ff4c121814aa5/Objects/frameobject.c#L1375
[2]: https://github.com/python/cpython/blo[]d1f15712d243e03e50332abb86ff4c121814aa5/Objects/frameobject.c#L1255-L1256

Failing tests: before=155, after=135

PiperOrigin-RevId: 657944178