Accept different errors due to duplicate return opcodes in 3.12
Like in 3.10, 3.12 compiles the following into 2 RETURN_VALUE opcodes for the same return statement:

    def foo(x: str) -> int:
      if x:
        y = "foo"
      else:
        y = 3j
      return y

This causes pytype to emit 2 different errors on that line rather than 1 combined error.

This is currently kinda working as designed. Since pytype generates types from the bytecode the exact error can change if the bytecode changes. For now we accept those differences.

Some tests exist to catch specific errors that only occur when the type is merged into a union. Change the code to preserve those errors. E.g. ensuring the line containing a ternary isn't the last line in the block tends to revert the bytecode to what it was in 3.11.

Failing tests: before=121, after=106

PiperOrigin-RevId: 658342902