Fix type narrowing in `case _ as foo` for 3.12
Type narrowing of the match variable for default match branches happens in one of the opcodes in the line. 3.11 compiles `case _ as foo` to:

    COPY       1
    POP_TOP
    STORE_FAST 2 (foo)

Pytype used (I assume somewhat arbitrarily) the opcode before STORE_FAST to do the type narrowing.

3.12 however compiles `case _ as foo` to:

    STORE_FAST 2 (foo)

That is, there is no opcode before STORE_FAST. STORE_FAST is the only opcode. This fix moves the type narrowing to the STORE_FAST opcode itself, which works both in 3.11 and 3.12.

Failing tests: before=105, after=102

PiperOrigin-RevId: 658766612