Get rid of the test `pytype_extensions.test_pytype_extensions` being run separately.
The following change after this one, of which is to delete the dead typegraph/cfg.py (was replaced by cpp implementation long time ago) is failing on this specific test due to extension module not being visible from the python-only test configuration. The test needs to run under build_utils.build_script to be built with the extension modules properly to be able to be used in tests.

Before this change, it seems accidental that the tests are using the old code(cfg.py) and still passing.

PiperOrigin-RevId: 672864348