Pin versions of all dependencies for CI
This guards us from getting broken at HEAD by e.g. an update to a new version of pylint. Instead we'll be upgrading them whenever we are ready for them.

Now the CI experience should be totally reproducible.

Importantly, even packages that *we* own (pycnite, importlab) are pinned. So it is important that we do not forget to refresh dependencies whenever we make new releases of those.

PiperOrigin-RevId: 678337535