Sort blocks by line number of the first opcode
Blocks are sorted by number of predecessors. Previously the index of the first opcode, i.e. the order in the bytecode, served as a tie breaker.

In Python 3.12 the added test case fails with:

    dummy_input_file:10:3: error: in <module>: None [assert-type]
      Expected: Optional[str]
        Actual: None

      assert_type(err, str|None)
      ~~~~~~~~~~~~~~~~~~~~~~~~~~

Python 3.12 emits the bytecode for the `except` block last for some reason.

Because of the while loop, every block is a predecessor of every other block, and the predecessor-based sorting fails. It falls back to the tie breaker and pytype executes the `except` block last, i.e. after `assert_type`. By the time pytype executes `assert_type` it don't know about the possible `str` type.

This change uses the line number of the first opcode as a tie breaker first and then falls back to the index.

PiperOrigin-RevId: 679166584