Make `assert_type` more resilient with a string argument
Background
----------

Consider this example:

    assert_type(a, 'List[int]')

Pytype's historic behavior is to assert that its pretty-printed type output exactly matches the string. Also notably, it doesn't even require the types to exist in the current namespace.
Meanwhile, all other type checkers' behavior is to just error out, a string is not supposed to be passed. Users of pytype should also stop using strings if they are still doing so.

Because I want to change this pretty-printed output to become `'list[int]'`, anyone who had made such an assertion will now get a failure, because the string does not match anymore.

The change
----------

In case the string match doesn't succeed, I am making pytype fall back to the behavior as if this piece of code was:

     assert_type(a, List[int])

Which can in many cases keep these asserts working. Though it will still fail if `List` is not imported in the current module.

PiperOrigin-RevId: 683246569