Remove `Never` annotations when merging pyi files.
`Never` is a valid type annotation, but it's most likely wrong if it's inferred by pytype, and it has a chain effect that all downstream code starts to get treated as unreachable.

Fixing pytype to not infer `Never` on return types should be better, but this should be dealt as a separate fix.

PiperOrigin-RevId: 684355324