Add super-minimal support for `enum.member()`
This implementation is just as a no-op pass-through. This wrapper is not *actually* supposed to be no-op, but this is a close enough approximation. PyType already treats every member of an enum as a member. Wrapping it into `enum.member()` only makes it explicit and gives you the ability to write code that is compatible with both PyType and Python 3.13, which *doesn't* actually make everything a member anymore.

PiperOrigin-RevId: 820345768